lab_09/requirements.txt
aiohttp==3.8.1
aiohttp-apispec==2.2.1
aioredis==2.0.1
aiosignal==1.2.0
apispec==3.3.2
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
charset-normalizer==2.0.10
click==8.0.3
frozenlist==1.2.0
hiredis==2.0.0
idna==3.3
Jinja2==2.11.3
MarkupSafe==2.0.1
marshmallow==3.14.1
multidict==5.2.0
PyYAML==6.0
typing-extensions==4.0.1
webargs==5.5.3
yarl==1.7.2

matplotlib~=3.5.1